refactor(homepage): move files
