+++ src/components/homepage/Waves.tsx
@@ -1,0 +1,28 @@
+export const Waves = () => (
+	<div className="absolute bottom-0 w-full">
+		<svg
+			className="object-cover w-full"
+			xmlns="http://www.w3.org/2000/svg"
+			preserveAspectRatio="none"
+			viewBox="0 0 1920 200"
+		>
+			<path
+				width="100%"
+				d="M -0 100.045 L 64 100.045 C 128 100.045 256 100.045 384 84.333 C 512 68.326 640 37.345 768 29.231 C 896 21.117 1024 37.345 1152 52.835 C 1280 68.326 1408 84.554 1536 80.35 C 1664 76.44 1792 52.835 1856 41.033 L 1920 29.231 L 1920 218.068 L -0 218.068 Z"
+				opacity="0.49"
+				className="dark:fill-black fill-white"
+			></path>
+			<path
+				width="100%"
+				d="M 0 49.635 L 64 68.294 C 128 87.303 256 124.094 384 119.714 C 512 115.334 640 68.031 768 40.262 C 896 11.968 1024 3.208 1152 12.23 C 1280 21.603 1408 49.635 1536 68.294 C 1664 87.303 1792 96.063 1856 101.056 L 1920 105.698 L 1920 217.825 L 0 217.825 Z"
+				opacity="0.31"
+				className="dark:fill-black fill-white"
+			></path>
+			<path
+				width="100%"
+				d="M 0 84.4 L 64 89.882 C 128 95.778 256 106.122 384 122.982 C 512 139.222 640 161.978 768 145.118 C 896 128.878 1024 73.022 1152 73.332 C 1280 73.022 1408 128.878 1536 128.568 C 1664 128.878 1792 73.022 1856 45.818 L 1920 18.2 L 1920 216.8 L 0 216.8 Z"
+				className="dark:fill-black fill-white"
+			></path>
+		</svg>
+	</div>
+);
