+++ src/utils/formatNumbers.ts
@@ -1,0 +1,1 @@
+export const FormatThousands = (x: number): string => Math.floor(x / 100) / 10.0 + "k";
